Added blur in Algorithm
The blur is working. The blur can be turned on or off in painter.cs in
rule 43. I don't know how to make a button in the interface for this
yet.

Many other techniques than blurring are possible and really easy. The
blurring slows the proces down with half the speed.

I added the AForge references (three .dll files needed) in the folder
Algorithm\References, this should work automatically.

If AForge does not work, you have to download AForge from:
http://www.aforgenet.com/framework/downloads.html. However, you only
have to download the library and substract three .dll files:
(AForge.dll, AForgeImaging.dll, AForgeImaging.Formats.dll) and make a
correct reference in Visual studio in: Algorithm -> references -> (right
click) Add Reference and choose your own folder here for the .dll files.